feat: create separated folder for page wrapper
