+++ src/pages/ConnectDevicePage/ConnectDevicePage.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import BreadcrumbBar from '../../components/BreadcrumbBar'
 import { Button as StatusButton, Tag, Text, Avatar, Checkbox } from '@status-im/components'
 import { Label, Separator, XStack, YStack } from 'tamagui'
-import PageWrapperShadow from '../../components/PageWrapperShadow'
+import PageWrapperShadow from '../../components/PageWrappers/PageWrapperShadow'
 import NimbusLogo from '../../components/Logos/NimbusLogo'
 import Titles from '../../components/Titles'
 import NodeIcon from '../../components/Icons/NodeIcon'
